requirements.txt
annotated-types==0.7.0
anyio==4.4.0
astor==0.8.1
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
Cython==3.0.11
decorator==5.1.1
exceptiongroup==1.2.2
fastapi==0.112.1
fire==0.6.0
fonttools==4.53.1
fuzzywuzzy==0.18.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
imageio==2.35.1
imgaug==0.4.0
importlib-resources==6.4.4
kiwisolver==1.4.5
lazy-loader==0.4
lmdb==1.5.1
lxml==5.3.0
matplotlib==3.7.5
networkx==3.1
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt-einsum==3.3.0
packaging==24.1
paddleocr==2.8.1
paddlepaddle==2.6.1
pillow==10.4.0
protobuf<=3.20.2
pyclipper==1.3.0.post5
pydantic==2.8.2
pydantic-core==2.20.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-multipart==0.0.9
PyWavelets==1.4.1
PyYAML==6.0.2
rapidfuzz==3.9.6
requests==2.32.3
scikit-image==0.21.0
scipy==1.10.1
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.38.2
termcolor==2.4.0
tifffile==2023.7.10
tqdm==4.66.5
typing-extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
zipp==3.20.0